refac: Querying, Updating DAO를 통합 및 패키지 변경
