+++ src/main/java/nextstep/dao/WebReservationDAO.java
@@ -1,9 +1,14 @@
-package nextstep.dao.web;
+package nextstep.dao;
 
+import java.sql.Date;
+import java.sql.PreparedStatement;
+import java.sql.Time;
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
@@ -11,7 +16,7 @@ import java.time.LocalTime;
 import java.util.List;
 
 @Repository
-public class ReservationQueryingDAO {
+public class WebReservationDAO {
     private final RowMapper<Reservation> reservationRowMapper = (resultSet, rowNum) -> {
         Reservation reservation = new Reservation(
                 resultSet.getLong("id"),
@@ -29,8 +34,26 @@ public class ReservationQueryingDAO {
 
     private JdbcTemplate jdbcTemplate;
 
-    public ReservationQueryingDAO(JdbcTemplate jdbcTemplate) {
+    public WebReservationDAO(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Long insertWithKeyHolder(Reservation reservation) {
+        String sql = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?);";
+
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setDate(1, Date.valueOf(reservation.getDate()));
+            ps.setTime(2, Time.valueOf(reservation.getTime()));
+            ps.setString(3, reservation.getName());
+            ps.setString(4, reservation.getTheme().getName());
+            ps.setString(5, reservation.getTheme().getDesc());
+            ps.setInt(6, reservation.getTheme().getPrice());
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
     }
 
     public Reservation findReservationById(Long id) {
@@ -41,5 +64,10 @@ public class ReservationQueryingDAO {
     public List<Reservation> findReservationByDateAndTime(LocalDate localDate, LocalTime localTime) {
         String sql = "select * from reservation where date = ? and time = ?";
         return jdbcTemplate.query(sql, reservationRowMapper, localDate, localTime);
+    }
+
+    public int delete(Long id) {
+        String sql = "delete from reservation where id = ?";
+        return jdbcTemplate.update(sql, Long.valueOf(id));
     }
 }
