+++ src/test/java/nextstep/web/WebReservationDAOTest.java
@@ -1,7 +1,8 @@
 package nextstep.web;
 
-import nextstep.dao.web.ReservationQueryingDAO;
+import nextstep.dao.WebReservationDAO;
 import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,15 +19,15 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @JdbcTest
-class ReservationQueryingDAOTest {
-    private ReservationQueryingDAO reservationQueryingDAO;
+class WebReservationDAOTest {
+    private WebReservationDAO webReservationDAO;
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
     @BeforeEach
     void setUp() {
-        reservationQueryingDAO = new ReservationQueryingDAO(jdbcTemplate);
+        webReservationDAO = new WebReservationDAO(jdbcTemplate);
 
         List<Object[]> reservations = List.of(
                 new Object[]{Date.valueOf("2022-08-11"), Time.valueOf("13:00:00"), "name", "워너고홈", "병맛 어드벤처 회사 코믹물",
@@ -43,7 +44,7 @@ class ReservationQueryingDAOTest {
     @Test
     @DisplayName("ID를 통해 예약 정보를 잘 받아오는지 테스트")
     void findReservationByIdTest() {
-        Reservation reservation = reservationQueryingDAO.findReservationById(1L);
+        Reservation reservation = webReservationDAO.findReservationById(1L);
 
         assertThat(reservation).isNotNull();
         assertThat(reservation.getName()).isEqualTo("name");
@@ -52,12 +53,43 @@ class ReservationQueryingDAOTest {
     @Test
     @DisplayName("예약 날짜와 시간이 일치하는 예약 정보를 잘 받아오는지 테스트")
     void findReservationByDateAndTimeTest() {
-        List<Reservation> reservations = reservationQueryingDAO.findReservationByDateAndTime(
+        List<Reservation> reservations = webReservationDAO.findReservationByDateAndTime(
                 LocalDate.of(2022, 8, 11),
                 LocalTime.of(13, 0)
         );
 
         assertThat(reservations).hasSize(1);
         assertThat(reservations.get(0).getName()).isEqualTo("name");
+    }
+
+    @Test
+    @DisplayName("예약 정보를 생성하고 생성한 예약의 ID를 잘 받아오는지 테스트")
+    void insertWithKeyHolderTest() {
+        Reservation reservation = new Reservation(
+                LocalDate.of(2022, 8, 11),
+                LocalTime.of(13, 0),
+                "name3",
+                new Theme(
+                        "워너고홈",
+                        "병맛 어드벤처 회사 코믹물",
+                        29000
+                )
+        );
+        Long id = webReservationDAO.insertWithKeyHolder(reservation);
+
+        assertThat(id).isNotNull();
+
+        Reservation reservationById = webReservationDAO.findReservationById(id);
+        assertThat(reservationById).isNotNull();
+        assertThat(reservationById.getName()).isEqualTo("name3");
+    }
+
+
+    @Test
+    @DisplayName("ID가 일치하는 예약을 잘 삭제하는지 테스트")
+    void deleteTest() {
+        int rowNum = webReservationDAO.delete(1L);
+
+        assertThat(rowNum).isEqualTo(1);
     }
 }
